refactor: move modules & assets since intro folder is not necessary
