+++ src/entities/pixi-icon/assets/hanmogm.png
