+++ src/entities/pixi-icon/assets/hoshino.png
