+++ src/entities/pixi-icon/assets/javascript.png
