+++ src/entities/pixi-icon/assets/koishi.png
