+++ src/entities/pixi-icon/assets/react.png
