+++ src/entities/pixi-icon/assets/typescript.png
