+++ src/entities/pixi-icon/index.ts
